Refactor info field Move ontology term imports Close #240
